Full processing cycle + fake sensors

+++ app/src/domain/adc_domain/hardware/adc_config.h
@@ -5,7 +5,7 @@ namespace eerie_leap::domain::adc_domain::hardware {
 struct AdcConfig {
     int channel_count;
     int resolution;
-    int reference_voltage;
+    double reference_voltage;
 };
 
 }  // namespace eerie_leap::domain::adc_domain::hardware
